Untying the Gordian Knot of Version Following.
After some meditation we decided to leave the houdiniXX.X naming
convention, and bring the otls dir to the main level of qLib. You can
still put different versions of qLib into different houdiniXX.X dirs in
your system if you decide so.. We consider qLib as forward compatible
with houdini and will fix things if some of the assets ceases to work
with new versions. We won't test against legacy versions, but will
provide access to older versions of the lib through git.