Added 'parameter' folder.
This folder is intended to include common parameter patterns which can be
added to assets making it easier to create similar interfaces for similar
cases.

The graveyard_message.ds file includes the standard garveyard message that
has to be included before any other interface elements in assets belonging
to the graveyard section.