ENHANCED: Path Deform SOP.
WARNING: The defaults and the output of this operator have been changed
by this revision.

Now the original geometry's Z axis is deformed along the curve instead
of X.

Resize By Segment Length is now an option, and the original geometry can
keep its size. The later is the default.

Support of flat geometry like grids added.